Oops... I didn't know that we had this much stuff under utils, a lot of stuff was added when I wasn't looking :)
About your question, no, the defaults that were chosen when my patch was
applied were to silence most of the debugging output. And as I didn't
think about stuff outside the client and server directories (sorry about
my naiveness on that, I can prepare a patch covering other stuff I might
have left behind).

Bottom line is, when you run packager, it's not server, client or tests,
so the logging mechanism doesn't know how to setup the root logger for
it. The simple solution is just add the debug library, setup it for a
new 'utils' class, add 'utils' to the debug.ini file and then set up
'utils' to the DEBUG level. Here's a quick patch that does that.

From: Lucas Meneghel Rodrigues <lmr@linux.vnet.ibm.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@2294 592f7852-d20e-0410-864c-8624ca9c26a4