refix the downloadable test URL regular expression
Recently a change was made to fix an unbalanced parenthesis in the
package name decoder regular expression.  While this fixed the imbalance
it incorrectly added an additional bracketed section in that regexp
leading to a 'too many opjects in unpack'.  Fix this up.

With this patch applied I am able to use the downloadable test system again.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@2334 592f7852-d20e-0410-864c-8624ca9c26a4