Modified machine_aggr.cgi to accept benchmark and key values as a list of paired values instead of two lists.
machine_aggr.cgi accepts a list of benchmarks and a list of keys as two
seperate inputs. Within the code, this is zipped together into a list
of benchmark-key pairs.

This CL modifies machine_aggr.cgi to accept a list of benchmark-key
pairs from the user. Sample input:

        benchmark_key=benchmark1:key1,benchmark2:key2,...

To maintain backward compatibility, the old interface is left intact.
The user is expected to use exactly one of the two options. If both
interfaces are used, preference is given to the new interface.

From: Nikhil Rao <ncrao@google.com>
Signed-off-by: Martin Bligh <mbligh@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@2468 592f7852-d20e-0410-864c-8624ca9c26a4