Rewrite the get_file and send_file code in AbstractSSHHost to properly support rsync-style trailing slash paths.
Currently, the interface claims to support this but in order to
provide support for working with rsync and scp the functions map
trailing slash paths like path/ to path/*. Unfortunately this isn't
quite the same - it misses dotfiles.

This rewrite changes this rather dramatically; it discards the
approach of doing the same thing with both rsync and scp. For rsync
it now just uses trailing slashes directly, and then if you need to
fall back to scp it replaces the path with the appropriate glob
patterns (which requires some work to determine which ones are
usable).

Risk: High
Visibility: get_file and send_file using trailing-slash source
directories should now work properly (i.e the same as rsync).

Signed-off-by: John Admanski <jadmanski@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@2605 592f7852-d20e-0410-864c-8624ca9c26a4