When aborting host queue entries, completely ignore completed entries.  They get ignored eventually anyway, but if we include them here, we can end up erroneously rejecting abort requests for synchronous jobs.
This happens because entries that fail verify get their execution_subdir set to their hostname. Then, when we see the abort, we find only one entry with that execution tag while the sync_count is >1, so we think the user is aborting part of a synchronous execution and throw an exception.

Signed-off-by: Steve Howard <showard@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@2658 592f7852-d20e-0410-864c-8624ca9c26a4