This overkill of a "script" will fetch, build and install python libraries that autotest depends on.  Currently this can deal with setuptools, MySQLdb, and Django.  It can be expanded in the future as needed.
NOTE: This installs django 1.0.2.  Don't use this script for that
until after my next change which ports autotest to Django 1.0.2.

TESTING:
I'm running a frontend on a Ubuntu dapper host and it is using
the MySQLdb and Django packages installed locally in my
autotest/site-packages directory by this script rather than anything
installed in the system's /usr/lib/python2.4/site-packages.

Signed-off-by: Gregory Smith <gps@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@2748 592f7852-d20e-0410-864c-8624ca9c26a4