Create job status only if the job has a host or a meta_host.
This patch addresses the problem where a get_hosts query to autotest
returns an empty list. This can happen when a host that was in the
DB at some point of time is removed from the DB. As a thumb rule, we
only create job status entries for job statuses that have either a
valid host or meta_host.

Signed-off-by: Nikhil Rao <ncrao@google.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@2837 592f7852-d20e-0410-864c-8624ca9c26a4