Hi Team,
I wrote some methods that extends the test_config.py. These methods are
to check the parameters in a config file. 

Today we did not had this functionality in autotest and need write some
functions inside our test to check if the test got a valid ip address
from config file or if the parameter got was a string or int.

To use this proposal, we will use a new format to describe the
parameters, this format is type.parameter, example:

ib.server = 192.168.1.1
str.password = mypass
int.mtu = 2048

This method can be extended to support new types, helping in the test
development, because the developer only needs use the method
test_cfg.check() to validate the config file, example:

test_cfg.check('test config section')

Regards,


git-svn-id: http://test.kernel.org/svn/autotest/trunk@3155 592f7852-d20e-0410-864c-8624ca9c26a4