KVM test: Do a better job at handling kvm module failures
Developers might want to use the KVM test to verify changes
only on their own local qemu source directory and will use the
'localsrc' install mode on the KVM test control file, and
provide 'load_modules': 'no' option on the params dict inside
the kvm control file. However, that option is not documented and
defaults to 'yes',  and when the code to load kvm modules can't
find a built kvm.ko module inside the source dir, an unhandled
error exception (totally unhelpful) will be thrown.

This patch makes the test throw a much more helpful error message,
helping the developer to figure out what is wrong and fix the
problem.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@3412 592f7852-d20e-0410-864c-8624ca9c26a4