Some fixes to history handling of label and attribute filters in CommonPanel.  The separation between state and view wasn't being properly handled.
I think the whole state/view refactoring is moving slowly towards this model-view-presenter thingy I saw at Google I/O.  When the need arises we should go ahead and make that a clean cut to avoid issues like this.  Increased open-in-new-window support will probably be the feature to drive that change (it was the motivation behind the original separation).

Signed-off-by: Steve Howard <showard@google.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@3460 592f7852-d20e-0410-864c-8624ca9c26a4