KVM test: kvm_subprocess: don't immediately check /proc/$PID/cmdline
kvm_spawn.is_alive() incorrectly assumes that /proc/$PID/cmdline immediately
reflects the new command line of the forked process.  This makes it report
false negatives occasionally.  To prevent that, perform the command line check
only if the process is more than 10 seconds old, or if the process has been
pickled and unpickled.  In both cases, there is more than enough time for
/proc/$PID/cmdline to get updated.

Signed-off-by: Michael Goldish <mgoldish@redhat.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@3513 592f7852-d20e-0410-864c-8624ca9c26a4