KVM test: kvm_subprocess: send user specified parameters to kvm_tail callbacks
(A mistake was made in the first version of this patch -- this is a corrected
version.)

Allow the user to specify parameters to send to output_func and
termination_func in addition to the regular output string and exit status.
The user specified parameters will be pre-pended to the other parameters.
This is mainly meant to bypass Python's refusal to pickle instance methods.

Signed-off-by: Michael Goldish <mgoldish@redhat.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@3534 592f7852-d20e-0410-864c-8624ca9c26a4