KVM test: allow the user to specify the paths of the qemu and qemu-img binaries
Currently, qemu and qemu-img are accessed via symlinks that must be present in
test.bindir (the KVM test dir).

This patch adds new parameters qemu_binary and qemu_img_binary, which specify
the paths of the qemu and qemu-img binaries.
They may be absolute paths or relative to test.bindir (the KVM test dir).
In kvm_tests.cfg.sample they are set to 'qemu' and 'qemu-img', which means the
binaries are expected to be found in test.bindir, which is the current behavior.

Adding these parameters results in slightly cleaner code, but also allows for
some more flexibility in defining tests.  For example, the user can 'variant'
on the parameter qemu_binary, i.e. define several variants of a test set that

differ only in this parameter.  This will make the test set run several times,
each time using a different pre-installed version of qemu.  The parameters also
make it possible to use pre-installed qemu and qemu-img that reside somewhere
outside the Autotest dir, e.g. in /usr/bin.

Signed-off-by: Michael Goldish <mgoldish@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@3550 592f7852-d20e-0410-864c-8624ca9c26a4