KVM test: Step files to setup the new remote shell server on Windows guests
This patch adds step files for Win2008.32, Win2008.64, WinVista.32,
WinVista.64, Win2003.32, Win2003.64, WinXP.32, WinXP.64 and Win2000.

I tried to make them as short as possible in order to minimize the chance of
failure.

This patch also adds the missing shutdown procedure to the end of the
Win2008.64 installation step file.

All files have been tested and seem to work properly, except Win2008.64, which
enters a reboot loop during installation (with several KVM versions).
This is not a step file problem, but the step file may contain errors because
it couldn't be properly tested.

Signed-off-by: Michael Goldish <mgoldish@redhat.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@3555 592f7852-d20e-0410-864c-8624ca9c26a4