Refactor code related to double list selecting widgets used in both SpreadsheetView and TableView.  This will facilitate code reuse and extension for the feature to include test iterations/attributes in table view.  It will also make unit testing possible and provides cleaner organization of the code.
The major effort here is to refactor the DoubleListSelector and HeaderSelect classes according to the Passive View pattern.  There are passive view classes for the double list itself (DoubleListSelector, which will be renamed in a future CL to DoubleListSelectDisplay) and for the spreadsheet header selection UI (HeaderSelectorView, which I think I'll also rename to HeaderSelectorDisplay).  There is a new class, MultiListSelectPresenter, which incorporates the presentation logic from DoubleListSelector as well as some of the logic from HeaderSelect related to switch between a single and double list.  The remaining code in HeaderSelect was modified to use a MultiListSelectPresenter as much as possible, retaining only the code specific to the spreadsheet header selection (and the management of machine label headers will likely be extracted and generalized too, since that sort of thing will be necessary for test attributes/iterations).  Finally, small modifications to Spreadsheet
 View and TableView were required to have them use the new interfaces.

This change also introduces the ToggleLink widget, consolidating logic that had been applied in multiple places previously.

Signed-off-by: Steve Howard <showard@google.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@3667 592f7852-d20e-0410-864c-8624ca9c26a4