KVM test: AutoIt and Autotest wrappers: use get_command_output() instead of sendline()
get_command_output() is safer because it waits for the prompt to return.
sendline() returns immediately, and the output generated in response to the
command can appear later and interfere with the test.
For example, the prompt can appear while the Autotest wrapper waits for an
Autotest test to complete, and this will make the wrapper exit as if the test
has completed.

Signed-off-by: Michael Goldish <mgoldish@redhat.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@3794 592f7852-d20e-0410-864c-8624ca9c26a4