KVM test: step file tests: add parameter timeout_multiplier
This parameter multiplies the timeout values of all the barriers in a step
file test. It is useful for slower hosts, under load (e.g. when executing
multiple tests in parallel) and for testing QEMU without KVM.  In any of
these cases, the multiplier should be greater than 1 in order to give the
test more time to complete.

In addition to modifying tests/steps.py, this patch adds a usage example to
control and control.parallel.

Signed-off-by: Michael Goldish <mgoldish@redhat.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@3796 592f7852-d20e-0410-864c-8624ca9c26a4