KVM test: Utility to make windows guests ACK end of install
In order to make windows guests to tell the unattended install
server that the install process ended, a simple C++ program
using the winsock 2 API had to be developed. The program was
based on an MSDN article. Its usage is

finish.exe [host-ip]

The test expect this to be compiled under deps. This can be
accomplished on linux hosts that support the mingw cross
compiling environment (or you can just compile it under your
windows guest). For Fedora 11, the instructions are:

yum install mingw32-w32api mingw32-gcc-c++

i686-pc-mingw32-g++ finish.cpp -lws2_32 -o finish.exe

Binaries produced by this process work just fine on my tests.
I still haven't made my mind about shipping binaries, with
the test, so in the meantime you want to check this compile
the code as you wish and put it under the same deps dir.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@3812 592f7852-d20e-0410-864c-8624ca9c26a4