KVM test: add option to kill all unresponsive VMs at the end of each test
This is useful for tests that may leave VMs in a bad state but can't afford to
use kill_vm_on_error = yes.
For example, timedrift.with_reboot can fail because the reboot failed or
because the time drift was too large.  In the latter case there's no reason to
kill the VM.

Signed-off-by: Michael Goldish <mgoldish@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@3837 592f7852-d20e-0410-864c-8624ca9c26a4