KVM test: Change the way subtests are loaded
Recently autoserv changed its default behavior of rsyncing
the whole client directory to the test machine, now it
will copy only the needed tests to the client machine.

Also, the way the tests are loaded when running from the
server has changed, breaking the KVM test when ran from
autoserv.

So change the mechanism to load KVM subtests, in order to
cope with the recent autoserv changes.

Thanks to Ryan Harper for having noticed this issue.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@3891 592f7852-d20e-0410-864c-8624ca9c26a4