Fix kvm_config.py -f <filename> mode
kvm_config.py supports specifying a different filename for
test config.  This patch fixes the option parsing parameters.
Currently it uses 'store_true' which stores the value True into
the filename variable; we want the 'store' mode which will store
the value of the option (aka, the filename) in the variable.

Signed-off-by: Ryan Harper <ryanh@us.ibm.com>



git-svn-id: http://test.kernel.org/svn/autotest/trunk@3894 592f7852-d20e-0410-864c-8624ca9c26a4